Bug fixes found and partially tested during deployment to production.
Adding here for discussion, will need testing in both cases. Specifically, install and launch a workflow with variables in the template that remain blank when launching. Also launch a new Java workflow using both wait and non-wait and make sure that no new workflow run objects are created and that the existing one is updated properly.


Former-commit-id: a48e78ed152c8a5e32172328176b955be30cd57f