Working integration tests that cover the user tutorial in order as well as some of the developer tutorial and admin tutorial.
Total runtime is 73 minutes, 16 seconds, also requires roughly 10GB of space. Some of this is in /tmp and has to be cleaned in the next iteration.
Run via the steps in the README, make sure to correctly set your .seqware/settings to hit the Jetty webservice during tests and turn off any cronjob entries relating to SeqWare.


Former-commit-id: 97d0b7b2178d41cca9c0f1531861fa5613dfed08